requirements.txt
asgiref==3.8.1
astroid==3.3.6
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
decouple==0.0.7
defusedxml==0.8.0rc2
dill==0.3.9
dj-database-url==3.0.0
Django==5.2.2
django-admin-logs==1.3.0
django-allauth==64.1.0
django-environ==0.12.0
django-extensions==4.1
django-widget-tweaks==1.5.0
Faker==37.5.3
mercadopago==2.3.0
gunicorn==23.0.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
pip==23.2.1
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.9.0
pylint==3.3.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==65.5.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==10.7.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2